requirements.txt
Flask
Flask-Cors
Flask-JWT-Extended
flask-marshmallow
Flask-SQLAlchemy
marshmallow-sqlalchemy
SQLAlchemy
Werkzeug
psycopg[binary]
python-dotenv
